+++ Components/FirstTimeSetup.lua
@@ -1,0 +1,72 @@
+local root = script.Parent.Parent
+
+local PluginModules = root:FindFirstChild("PluginModules")
+local Style = require(PluginModules:FindFirstChild("Style"))
+
+local includes = root:FindFirstChild("includes")
+local Roact = require(includes:FindFirstChild("Roact"))
+
+local Components = root:FindFirstChild("Components")
+local Button = require(Components:FindFirstChild("Button"))
+local ConnectTheme = require(Components:FindFirstChild("ConnectTheme"))
+
+local StandardComponents = require(Components:FindFirstChild("StandardComponents"))
+local StandardTextLabel = StandardComponents.TextLabel
+local StandardUIPadding = StandardComponents.UIPadding
+
+---
+
+--[[
+    props
+
+        onConfirm: () -> nil
+]]
+local FirstTimeSetup = Roact.PureComponent:extend("FirstTimeSetup")
+
+FirstTimeSetup.render = function(self)
+    local theme = self.props.theme
+
+    return Roact.createElement("Frame", {
+        AnchorPoint = Vector2.new(0.5, 0.5),
+        Size = UDim2.new(1, 0, 1, 0),
+        Position = UDim2.new(0.5, 0, 0.5, 0),
+        BackgroundTransparency = 0,
+        BorderSizePixel = 0,
+
+        BackgroundColor3 = theme:GetColor(Enum.StudioStyleGuideColor.MainBackground)
+    }, {
+        UIPadding = Roact.createElement(StandardUIPadding, {Style.PagePadding}),
+
+        Notice = Roact.createElement(StandardTextLabel, {
+            AnchorPoint = Vector2.new(0.5, 0),
+            Size = UDim2.new(1, 0, 0.5, 0),
+            Position = UDim2.new(0.5, 0, 0, 0),
+
+            Text = "Before using ColorPane for the first time, you will be prompted to allow script injection. This permission is necessary for other plugins to use the ColorPane API.",
+            TextXAlignment = Enum.TextXAlignment.Center,
+            TextYAlignment = Enum.TextYAlignment.Bottom,
+            TextWrapped = true,
+        }),
+
+        ConfirmButton = Roact.createElement(Button, {
+            AnchorPoint = Vector2.new(0.5, 0),
+            Position = UDim2.new(0.5, 0, 0.5, Style.SpaciousElementPadding),
+            Size = UDim2.new(0, 90, 0, Style.StandardButtonSize),
+
+            displayType = "text",
+            text = "I understand",
+
+            backgroundColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButton),
+            borderColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogButtonBorder),
+            hoverColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButton, Enum.StudioStyleGuideModifier.Hover),
+            displayColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButtonText),
+
+            disabledBackgroundColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButton, Enum.StudioStyleGuideModifier.Disabled),
+            disabledDisplayColor = theme:GetColor(Enum.StudioStyleGuideColor.DialogMainButtonText, Enum.StudioStyleGuideModifier.Disabled),
+
+            onActivated = self.props.onConfirm,
+        }),
+    })
+end
+
+return ConnectTheme(FirstTimeSetup)